Bug 444481 - Don't unmap the vDSO on s390x
Newer Linux kernels on s390x may use the vDSO as a "trampoline" for
syscall restart.  This means that the vDSO is no longer optional, and
unmapping it may lead to a segmentation fault when a system call restart
is performed.

So far Valgrind has been unmapping the vDSO on s390x.  Just don't do this
anymore.