Powerpc test_isa_3_1_VRT fix
The vcmpequq, vcmpgtsq, vcmpgtuq test cases for the dotted versions of the
instructions were issuing the non-dotted instruction.

This patch fixes the issues and updates the expected output.

Note, the issue exposed a bug in the VEX/priv/guest_ppc_toIR.c handling of
the instructions.  That fix is in a separate patch.