Bug 465782 - s390x: Drop -mzarch -march=z900 from assembler options
The -mzarch flag is unsupported by Clang, and it is redundant on 64-bit
build systems.  Remove it.

Also remove '-march=z900', since it is unsupported by Clang as well.  It
would only be needed on build systems with a default architecture lower
than z900.  Such systems are out of service for some time now.