Bug 444110 priv/guest_ppc_toIR.c: warning: duplicated 'if' condition
The compiler reported a duplicated condition in VEX/priv/guest_ppc_toIR.c
The handling of the plbz and xxpermx instructions have the same if/elseif
conditions.  The else if condition for the plbz instruction was wrong.  The
elseif statement should be checking for pType2 not pType1.  The plbz
instruction was inadvertently being handled by the else statement for
the lbz instruction.

This patch fixes the checking for the plbz and lbz instructions.