s390x: Replace absolute jump for a relative one
The bne instruction expects an absolute target address and it isn't
best-suited for implementing a short range jump, such as the one in
XCHG_M_R().
Replace it with jne which expects a relative address that can be
correctly computed a link time.

Interestingly, the jump is almost never taken. If it would, this would
crash the test.  However, linkers may complain when relacating the
target address used in bne.