Helgrind: add suppression for libnss from getaddrinfo
Seems as though Raspberry Pi OS doesn't have a debuginfo package
for libnss_mdns4 so we need to suppress based on the obj name.

There is still one error in the Helgrind getaddrinfo test, but that
is in the somewhat malformed libarmmem library (which does have
debuginfo but Valgrind can't read it because there is no ELF type
or size info). I think that suppressing that would be too broad.