s390x: Re-implement `vpdi' without intermediate updates
The vpdi instruction is implemented incorrectly for the case where the
result vector coincides with one of the input operands, because it
performs an intermediate update.  An example where this produces wrong
results is for

	vpdi    %v3,%v3,%v3,4

where the instruction is used to swap the two halves of %v3.

Fix this by avoiding the intermediate update and using Iop_64HLtoV128
instead.